+++ scripts/delayed.js
@@ -1,9 +1,13 @@
 // eslint-disable-next-line import/no-cycle
 import { sampleRUM, loadScript } from './lib-franklin.js';
 // eslint-disable-next-line import/no-cycle
-import { loadConsentManager } from './scripts.js';
+import { loadConsentManager, loadhubspotform } from './scripts.js';
 // Core Web Vitals RUM collection
 sampleRUM('cwv');
+// Loading hubspot form
+await loadhubspotform();
+// Script for Cookie consent manager
+await loadConsentManager();
 
 // google tag manager
 const gtmId = 'GTM-WSGNHXL';
@@ -28,21 +32,38 @@ console.log("TrustArc Events Binding..."); var dispatched = {}; var i = self.pos
 // add more delayed functionality here
 
 // Importing script for Hubspot form implementation in contact-us page
-const contactFormBlock = document.querySelector('.block.contact-form');
-if (contactFormBlock) {
-  const script = document.createElement('script');
-  script.setAttribute('type', 'text/javascript');
-  script.src = 'https://js.hsforms.net/forms/embed/v2.js';
-  script.addEventListener('load', () => {
-    // eslint-disable-next-line no-undef
-    hbspt.forms.create({
-      region: 'na1',
-      portalId: '14560092',
-      formId: 'af288827-9574-4a29-89eb-1b853eae1c20',
-    });
-  });
-  contactFormBlock.append(script);
-}
+// const contactFormBlock = document.querySelector('.block.contact-form');
+// if (contactFormBlock) {
+//   const script = document.createElement('script');
+//   script.setAttribute('type', 'text/javascript');
+//   script.src = 'https://js.hsforms.net/forms/embed/v2.js';
+//   script.addEventListener('load', () => {
+//     // eslint-disable-next-line no-undef
+//     hbspt.forms.create({
+//       region: 'na1',
+//       portalId: '14560092',
+//       formId: 'af288827-9574-4a29-89eb-1b853eae1c20',
+//     });
+//   });
+//   contactFormBlock.append(script);
+// }
 
-// Script for Cookie consent manager
-await loadConsentManager();
+// async function loadhubspotform() {
+//   const contactFormBlock = document.querySelector('.block.contact-form');
+//   if (contactFormBlock) {
+//     const script = document.createElement('script');
+//     script.setAttribute('type', 'text/javascript');
+//     script.src = 'https://js.hsforms.net/forms/embed/v2.js';
+//     script.addEventListener('load', () => {
+//       // eslint-disable-next-line no-undef
+//       hbspt.forms.create({
+//         region: 'na1',
+//         portalId: '14560092',
+//         formId: 'af288827-9574-4a29-89eb-1b853eae1c20',
+//       });
+//     });
+//     contactFormBlock.append(script);
+//   }
+// }
+
+// await loadhubspotform();
